Set serial_port to "auto"
With an upcoming change in slider_box ([PR
2](https://github.com/open-dynamic-robot-initiative/slider_box/pull/2))
the behaviour changes such that auto-detection is only done if the port
is explicitly set to "auto".  Thus changing it here to maintain the
current behaviour.

This change is backwards compatible with older versions of slider_box as
there auto-detection is done in any case.